make transactions negative and deposits positive

+++ project/main/views.py
@@ -42,56 +42,59 @@ def create_and_update_accounts(response, item):
 def create_and_update_transactions(response):
     for transaction in response.get('transactions', []):
         tid = transaction['transaction_id']
-        tran = Transaction.objects.filter(pk=tid).first() or Transaction(transaction_id=tid)
-        tran.account = Account.objects.filter(pk=transaction['account_id']).first()
-        tran.amount = transaction['amount']
-        tran.name = transaction['name']
-        tran.pending = transaction['pending']
-        tran.date = datetime.datetime.strptime(transaction['date'], '%Y-%m-%d').date()
+        account_obj = Account.objects.filter(pk=transaction['account_id']).first()
+        if account_obj.track:
+            tran = Transaction.objects.filter(pk=tid).first() or Transaction(transaction_id=tid)
+            tran.account = account_obj
+            tran.amount = transaction['amount']
+            tran.amount = tran.amount * -1
+            tran.name = transaction['name']
+            tran.pending = transaction['pending']
+            tran.date = datetime.datetime.strptime(transaction['date'], '%Y-%m-%d').date()
 
-        location = transaction.get('location', {})
-        loc = Location.objects.filter(
-                address=location['address'],
-                city=location['city'],
-                state=location['state'],
-                zip=location['zip'],
-                lat=location['lat'],
-                lon=location['lon'],
-                ).first()
-        if not loc:
-            loc = Location(
+            location = transaction.get('location', {})
+            loc = Location.objects.filter(
                     address=location['address'],
                     city=location['city'],
                     state=location['state'],
                     zip=location['zip'],
                     lat=location['lat'],
                     lon=location['lon'],
-                    )
-            loc.save()
+                    ).first()
+            if not loc:
+                loc = Location(
+                        address=location['address'],
+                        city=location['city'],
+                        state=location['state'],
+                        zip=location['zip'],
+                        lat=location['lat'],
+                        lon=location['lon'],
+                        )
+                loc.save()
 
-        tran.location = loc
-        tran.save()
+            tran.location = loc
+            tran.save()
 
-        categories = list(enumerate(transaction.get('category', [])))
-        for index, category in categories:
-            cat = Category.objects.filter(pk=category).first()
-            if not cat:
-                cat = Category(token=category)
-                parent = Category.objects.filter(pk=categories[index][1]).first()
-                if parent:
-                    cat.parent = parent
-                cat.save()
+            categories = list(enumerate(transaction.get('category', [])))
+            for index, category in categories:
+                cat = Category.objects.filter(pk=category).first()
+                if not cat:
+                    cat = Category(token=category)
+                    parent = Category.objects.filter(pk=categories[index][1]).first()
+                    if parent:
+                        cat.parent = parent
+                    cat.save()
 
-            existing_cat_rel = Transaction.objects.filter(
-                    pk=tran.pk,
-                    categories__pk=cat.pk)
+                existing_cat_rel = Transaction.objects.filter(
+                        pk=tran.pk,
+                        categories__pk=cat.pk)
 
-            if not existing_cat_rel:
-                tran.categories.add(cat)
+                if not existing_cat_rel:
+                    tran.categories.add(cat)
 
-        ptid = transaction['pending_transaction_id']
-        if ptid:
-            ptran = Transactions.object.get(pk=ptid).delete()
+            ptid = transaction['pending_transaction_id']
+            if ptid:
+                ptran = Transactions.object.get(pk=ptid).delete()
 
 # Views
 def index(request):
